adding images and tags to posts

+++ app/Http/Controllers/PostController.php
@@ -50,18 +50,51 @@ class PostController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'body' => 'required',
+            'tags' => 'required',
+            'cover_image' => 'image|nullable|max:1999',
         ]);
+
+        // Handle File Upload
+        if($request->hasFile('cover_image')){
+           // Get filename with the extension
+           $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+           // Get just filename
+           $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+           // Get just ext
+           $extension = $request->file('cover_image')->getClientOriginalExtension();
+           // Filename to store
+           $fileNameToStore= $filename.'_'.time().'.'.$extension;
+           // Upload Image
+           $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
+
+
 
         //create post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->tags = $request->input('tags');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success','Post Created');
     }
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function show($id)
+    {
+        $post = Post::find($id);
+        return view('posts.show')->with('post', $post);
 
+    }
 
     /**
      * Show the form for editing the specified resource.
@@ -72,6 +105,13 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if(!isset($post)){
+            return redirect('/posts')->with('error', 'No Post Found');
+        }
+
+        if(auth()->user()->id !== $post->user_id){
+            return reditrct('/posts')->with('error', 'Unauthorized Page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -91,8 +131,10 @@ class PostController extends Controller
 
         //edit post
         $post = Post::find($id);
+
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+
         $post->save();
 
         return redirect('/posts')->with('success','Post Updated');
@@ -107,10 +149,19 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
+
+
         $post->delete();
         return redirect('/posts')->with('success','Post Removed');
 
     }
-    
+
+
 
 }
